Feat: Implement continuation from selected choices
Handled page-flipping animation using the react-pageflip library and
implemented continuation from the selected point using the startPage prop.

Disabled the basic click events provided by react-pageflip by using the
useMouseEvents and disableFlipByClick props, and changed the onClick event
handling to the Image and Text components.
The reason is that the basic click events are prevPage and nextPage,
which did not work properly when moving to the next page
in a tree-structured storybook.

Implemented page turning using the onFlip event listener and a useEffect
that depends on curPage. By obtaining a ref to HTMLFlipBook using useRef
and using the bookRef.current.pageFlip().flip() function,
we can flip to the desired page with animation. Since the relationship
between curPage.page and the index in HTMLFlipBook is curPage.page * 2 - 2,
we utilized this to move the page whenever curPage changes.

Finally, there was an issue where if click events occurred multiple times,
the HTMLFlipBook page and curPage values would get out of sync.
To fix this, we managed the onChangeState event of HTMLFlipBook with
state and added a condition to execute the flip function
only when it is not "flipping".